thêm lightning ball

+++ Assets/Scripts/ScriptableObjects/MapObject/ShootableObject/EffectObjectConfig/LightningBallConfig.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "LightningBallConfig", menuName = "ScriptableObjects/ShootableObject/LightningBallConfig", order = 2)]
 public class LightningBallConfig : MapObjectConfig
 {
-    public float TARGET_COUNT = 4;
+    public int TARGET_COUNT = 4;
     public float RADIUS = 3;
     public float DELAY_BETWEEN_CHAIN = 0.05f;
 }
